separate vectorized Ewald class to new .py
