+++ src/les/module/__init__.py
@@ -4,6 +4,8 @@ from les.module.atomwise import *
 
 from les.module.ewald import *
 
+from les.module.ewald_vectorized import *
+
 from les.module.bec import *
 
 from les.module.fixedcharges import *
